Change temperature

+++ src/models.go
@@ -1,10 +1,10 @@
 package src
 
 type TextRequest struct {
-	Model       string `json:"model"`
-	Prompt      string `json:"prompt"`
-	Temperature int    `json:"temperature"`
-	MaxTokens   int    `json:"max_tokens"`
+	Model       string  `json:"model"`
+	Prompt      string  `json:"prompt"`
+	Temperature float32 `json:"temperature"`
+	MaxTokens   int     `json:"max_tokens"`
 }
 
 type TextResponse struct {
